duration is saved and showed

+++ app/src/main/java/com/powernote/project/powernote/model/Task.java
@@ -22,7 +22,7 @@ public class Task extends Note{
 
     public Task(){}
 
-    public double getDuration() {
+    public long getDuration() {
         return duration;
     }
